requirements.txt
canonicalwebteam.flask-base==1.0.5
alembic==1.7.5
canonicalwebteam.http==1.0.3
canonicalwebteam.blog==6.4.0
canonicalwebteam.search==1.2.1
canonicalwebteam.templatefinder==1.0.0
canonicalwebteam.image-template==1.3.1
canonicalwebteam.discourse==4.0.9
python-dateutil==2.8.2
pytz==2020.5
maxminddb-geolite2==2018.703
Flask-OpenID==1.3.0
feedgen==0.9.0
feedparser==6.0.8
pymacaroons==0.13.0
pybreaker==0.7.0
Flask-WTF==0.15.1
requests==2.27.0
sqlalchemy==1.3.23
marshmallow==3.14.1
mistune==0.8.4
psycopg2-binary==2.9.3
pyyaml==5.4.1
ubuntu-release-info==21.1
launchpadlib==1.10.15.1
macaroonbakery==1.3.1
sortedcontainers==2.4.0
vcrpy-unittest==0.1.7
webargs==7.0.1
jinja2==3.0.3
pillow
